apps/backend/requirements.txt
grpcio==1.62.1
grpcio-tools==1.62.1
protobuf==4.25.3
langchain>=0.2.0
langchain-openai>=0.1.0
langgraph>=0.2.0
langgraph-checkpoint>=1.0.0
langgraph-checkpoint-sqlite>=1.0.0
langgraph-checkpoint-postgres>=1.0.0
python-dotenv==1.0.1
uuid==1.30
langchain-core>=0.2.0
langchain-community>=0.2.0 # Added for ChatOllama
langchain-text-splitters>=0.2.0
fastapi==0.110.0
uvicorn[standard]==0.28.0
websockets>=12.0
jinja2==3.1.3
aiofiles==23.2.1
langchain-google-genai>=0.1.0
cryptography>=42.0.0
redis>=5.0.0
langchain-postgres>=0.0.1
psycopg2-binary>=2.9.0
opensearch-py>=2.4.0
python-multipart>=0.0.9
psutil>=5.9.0
ollama==0.1.7 # Added for local LLM integration
tortoise-orm>=0.20.0 # Added for ORM
asyncpg>=0.29.0 # PostgreSQL driver for Tortoise ORM
nest-asyncio>=1.6.0
